+++ src/client/java/net/detectivekaktus/client/data/item/primitive/DotcPrimitivesRecipeProvider.java
@@ -1,5 +1,6 @@
 package net.detectivekaktus.client.data.item.primitive;
 
+import net.detectivekaktus.item.tool.DotcTools;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricRecipeProvider;
 
@@ -168,6 +169,57 @@ public class DotcPrimitivesRecipeProvider extends FabricRecipeProvider {
                 .unlockedBy(
                         "has_radiant_crystal",
                         InventoryChangeTrigger.TriggerInstance.hasItems(DotcIngredients.RADIANT_CRYSTAL)
+                )
+                .save(exporter);
+
+        ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, DotcPrimitives.SANGE)
+                .pattern(" | ")
+                .pattern(" @ ")
+                .pattern(" # ")
+                .define('|', DotcIngredients.DIRE_CRYSTAL)
+                .define('@', DotcPrimitives.OGRE_AXE)
+                .define('#', DotcIngredients.BELT_OF_STRENGTH)
+                .unlockedBy(
+                        "has_ogre_axe",
+                        InventoryChangeTrigger.TriggerInstance.hasItems(DotcPrimitives.OGRE_AXE)
+                )
+                .save(exporter);
+
+        ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, DotcPrimitives.YASHA)
+                .pattern(" | ")
+                .pattern(" @ ")
+                .pattern(" # ")
+                .define('|', DotcIngredients.RADIANT_CRYSTAL)
+                .define('@', DotcPrimitives.BLADE_OF_ALACRITY)
+                .define('#', DotcIngredients.BAND_OF_ELVENSKIN)
+                .unlockedBy(
+                        "has_blade_of_alacrity",
+                        InventoryChangeTrigger.TriggerInstance.hasItems(DotcPrimitives.BLADE_OF_ALACRITY)
+                )
+                .save(exporter);
+
+        ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, DotcPrimitives.KAYA)
+                .pattern(" | ")
+                .pattern(" @ ")
+                .pattern(" # ")
+                .define('|', DotcIngredients.RADIANT_CRYSTAL)
+                .define('@', DotcPrimitives.STAFF_OF_WIZARDRY)
+                .define('#', DotcIngredients.ROBE_OF_THE_MAGI)
+                .unlockedBy(
+                        "has_staff_of_wizardry",
+                        InventoryChangeTrigger.TriggerInstance.hasItems(DotcPrimitives.STAFF_OF_WIZARDRY)
+                )
+                .save(exporter);
+
+        ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, DotcPrimitives.DRAGON_LANCE)
+                .pattern("@#@")
+                .pattern("@@@")
+                .pattern(" # ")
+                .define('#', DotcIngredients.MITHRIL_INGOT)
+                .define('@', DotcIngredients.DIRE_CRYSTAL)
+                .unlockedBy(
+                        "has_dire_crystal",
+                        InventoryChangeTrigger.TriggerInstance.hasItems(DotcIngredients.DIRE_CRYSTAL)
                 )
                 .save(exporter);
     }