+++ CHANGELOG.md
@@ -25,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Urn of Shadows item.
 * Spirit Soul Release effect.
 * Spirit Vessel item.
+* Force Staff sword item.
+* Dragon Lance sword item.
+* Hurricane Pike sword item.
 * Kill streak announcement which can be turned on and off in the mod configuration.